+++ weather-booking-app-client/src/components/BookWeatherComponents/Interface/WeatherHudProps.tsx
@@ -1,7 +1,7 @@
 import { TemperatureLevel } from "../../../endpoint-caller/interfaces/enums/TemperatureLevel";
 import { WindLevel } from "../../../endpoint-caller/interfaces/enums/WindLevel";
-import { BookingWeatherOption } from "../../../pages/BookingPage/BookingPage";
 import { BookingPageState } from "../../../pages/BookingPage/Interface/BookingPageState";
+import { BookingWeatherOption } from "../../../pages/BookingPage/Interface/BookingWeatherOptions";
 
 interface WeatherHudProps {
   weatherData: BookingPageState;
