+++ weather-booking-app-client/src/components/ConfirmBookingDetails/Interface/ConfirmBookingDetailsProps.tsx
@@ -1,6 +1,6 @@
 import { RouteComponentProps } from "react-router-dom";
 import { BookingPageState } from "../../../pages/BookingPage/Interface/BookingPageState";
-import { BookingWeatherOption } from "../../../pages/BookingPage/BookingPage";
+import { BookingWeatherOption } from "../../../pages/BookingPage/Interface/BookingWeatherOptions";
 
 interface ConfirmBookingDetailsProps extends RouteComponentProps {
   weatherBookingDetails: BookingPageState;
